3_crew/content/pyproject.toml
[project]
name = "content"
version = "0.1.0"
description = "content using crewAI"
authors = [{ name = "Your Name", email = "you@example.com" }]
requires-python = ">=3.10,<3.14"
dependencies = [
    "crewai[tools]>=0.152.0,<1.0.0",
    "openai>=1.30.0",
    "pillow>=10.2.0",
    "pydantic>=2.7.0",
    "python-slugify>=8.0.0"
]

[project.scripts]
content = "content.main:run"
run_crew = "content.main:run"
train = "content.main:train"
replay = "content.main:replay"
test = "content.main:test"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.crewai]
type = "crew"
